Change the type of the mHistoryAcceptRateMH into float to gain in memory. Reactivate the backup and Optimize the saving of the trace in the .dat file.

+++ src/mcmc/MHVariable.h
@@ -55,7 +55,7 @@ public:
     // On en affiche des sous-parties (correspondant aux chaines) dans la vue des résultats
     // A stocker dans les résultats!
     
-    QVector<double> mHistoryAcceptRateMH;
+    QVector<float> mHistoryAcceptRateMH;
     
     QString mProposal;
 };
